requirements.txt
Flask==1.1.2
numpy==1.18.4
Pillow==7.1.2
flasgger==0.8.1
tensorflow==2.1.0
Keras==2.3.1